refactor(agent-runtime): own the session record's history state
Opening a recorder, settling a write that just failed, and deciding that
recording may resume are the only three places that set a session's
`HistoryState`. Stopping and resuming are the same judgement read in opposite
directions, so they belong together rather than beside the operations that
happen to trigger them.

Moves them into agent_runtime/record.rs, which brings agent_runtime.rs from
1180 to 1076 lines. actor.rs, connection.rs and executor.rs fell below the
800-line gate earlier in this branch, so their exceptions are removed here
rather than lowered, leaving four reviewed exceptions where there were seven.
